z_iptables: Cleanup AuroraStore exceptions
The stable version of AuroraStore 4.x has again the same app name as 3.x:
namely 'com.aurora.store'. It is intended to directly update over 3.x
Therefore, the interim app names for the alpha and beta builds are not
needed any longer.

The exception for nighly builds however will remain, to allow users to
install the nightly next to the official stable build.

Change-Id: I081623202ab86c7544c5c174e75079b39ba624a7